+++ rakitaj/lox_interpreter/src/ast.rs
@@ -1,21 +1,18 @@
 // Questions
 // 1. How do I represent a tree of different types in rust?
 
-use crate::tokens::TokenType;
+use crate::tokens::Token;
 
 pub enum Expr {
-    Binary(Box<Expr>, TokenType, Box<Expr>),
+    Binary(Box<Expr>, Token, Box<Expr>),
     Grouping(Box<Expr>),
     LiteralBool(bool),
     LiteralNumber(f32),
     LiteralNull,
     LiteralString(String),
-    Unary(TokenType, Box<Expr>)
+    Unary(Token, Box<Expr>)
 }
 
 pub struct Ast {
-    node: Expr
+    pub node: Expr
 }
-
-
-
