+++ rakitaj/lox_interpreter/src/tokens.rs
@@ -1,3 +1,5 @@
+use std::fmt;
+
 #[derive(Debug, PartialEq)]
 #[derive(Clone)]
 pub enum TokenType {
@@ -48,4 +50,36 @@ pub enum TokenType {
     While,
 
     Eof,
+}
+
+impl fmt::Display for TokenType {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{:?}", self)
+    }
+}
+
+#[derive(Debug, PartialEq)]
+pub struct Token {
+    token_type: TokenType,
+    line: usize
+}
+
+impl Token {
+    pub fn new(
+        token_type: TokenType,
+        line: usize
+    ) -> Self {
+        Token {
+            token_type: token_type,
+            line: line
+        }
+    }
+}
+
+impl fmt::Display for Token {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(
+            f, "{} {}", &self.token_type, &self.line
+        )
+    }
 }
